+++ feo3boy/src/gbz80core/microcode/mod.rs
@@ -1,131 +1,15 @@
 use std::ops::Index;
 use std::slice::SliceIndex;
 
-#[cfg(feature = "microcode")]
 use once_cell::sync::Lazy;
 
-#[cfg(feature = "microcode")]
-use crate::gbz80core::oputils::{add8_flags, halt, rotate_left9, rotate_right9, sub8_flags};
-#[cfg(feature = "microcode")]
-use crate::gbz80core::CpuContext;
 use crate::gbz80core::Flags;
 
-/// Trait for the microcode stack.
-#[derive(Debug, Default, Clone, PartialEq, Eq)]
 #[cfg(feature = "microcode")]
-pub struct MicrocodeStack {
-    bytes: Vec<u8>,
-}
+mod r#impl;
 
 #[cfg(feature = "microcode")]
-impl MicrocodeStack {
-    /// Check if the microcode stack is empty.
-    fn is_empty(&self) -> bool {
-        self.bytes.is_empty()
-    }
-
-    /// Push a u8 onto the microcode stack.
-    fn pushu8(&mut self, val: u8) {
-        self.bytes.push(val);
-    }
-
-    /// Pop a u8 from the microcode stack.
-    fn popu8(&mut self) -> u8 {
-        self.bytes
-            .pop()
-            .expect("Cannot pop u8, microcode stack empty")
-    }
-
-    /// Read the top u8 without popping.
-    #[allow(unused)]
-    fn peeku8(&self) -> u8 {
-        *self
-            .bytes
-            .last()
-            .expect("Cannot peek u8, microcode stack empty")
-    }
-
-    /// Push a u8 onto the microcode stack.
-    fn pushu16(&mut self, val: u16) {
-        let val = val.to_le_bytes();
-        self.bytes.extend_from_slice(&val);
-    }
-
-    /// Pop a u16 from the microcode stack.
-    fn popu16(&mut self) -> u16 {
-        debug_assert!(
-            self.bytes.len() >= 2,
-            "Not enough bytes to pop u16, need 2, have {}",
-            self.bytes.len()
-        );
-        let mut val = [0u8; 2];
-        val.copy_from_slice(&self.bytes[self.bytes.len() - 2..]);
-        self.bytes.truncate(self.bytes.len() - 2);
-        u16::from_le_bytes(val)
-    }
-
-    /// Peek the top u8 without popping.
-    #[allow(unused)]
-    fn peeku16(&self) -> u16 {
-        debug_assert!(
-            self.bytes.len() >= 2,
-            "Not enough bytes to peek u16, need 2, have {}",
-            self.bytes.len()
-        );
-        let mut val = [0u8; 2];
-        val.copy_from_slice(&self.bytes[self.bytes.len() - 2..]);
-        u16::from_le_bytes(val)
-    }
-
-    /// Duplicates the specified number of bytes on the top of the stack.
-    fn ustack_dup(&mut self, count: usize) {
-        debug_assert!(
-            self.bytes.len() >= count,
-            "Not enough bytes on the microcode stack to duplicated {} bytes, only have {}",
-            count,
-            self.bytes.len()
-        );
-        self.bytes.extend_from_within(self.bytes.len() - count..);
-    }
-
-    /// Swaps `top_count` bytes on the top of the stack with `second_count` bytes below
-    /// it.
-    fn ustack_swap(&mut self, top_count: usize, second_count: usize) {
-        debug_assert!(
-            self.bytes.len() >= top_count + second_count,
-            "Not enough bytes on the microcode stack to swap {} bytes with {} bytes, only have {}",
-            top_count,
-            second_count,
-            self.bytes.len()
-        );
-        let second_start = self.bytes.len() - (top_count + second_count);
-        let second_end = self.bytes.len() - top_count;
-        self.bytes.extend_from_within(second_start..second_end);
-        self.bytes.drain(second_start..second_end);
-    }
-
-    /// Discards the specified number of bytes from the top of the stack.
-    fn ustack_discard(&mut self, count: usize) {
-        debug_assert!(
-            self.bytes.len() >= count,
-            "Not enough bytes on the microcode stack to discard {} bytes, only have {}",
-            count,
-            self.bytes.len()
-        );
-        self.bytes.drain(self.bytes.len() - count..);
-    }
-}
-
-/// Result of executing a microcode instruction.
-#[derive(Debug, Copy, Clone, Eq, PartialEq)]
-#[cfg(feature = "microcode")]
-pub enum MicrocodeFlow {
-    /// Microcode instruction requires a yield. CPU should stop executing microcode until
-    /// 4 ticks have elapsed.
-    Yield1m,
-    /// Continue executing microcode.
-    Continue,
-}
+pub use r#impl::{MicrocodeFlow, MicrocodeStack};
 
 /// Builder for microcode.
 #[derive(Default, Debug, Clone, PartialEq, Eq)]
@@ -190,6 +74,7 @@ impl MicrocodeBuilder {
     }
 
     /// Pop a single byte boolean off the microcode stack and run this if it is false.
+    #[allow(unused)]
     pub fn if_false(code_if_false: impl Into<MicrocodeBuilder>) -> Self {
         MicrocodeBuilder::cond(MicrocodeBuilder::new(), code_if_false)
     }
@@ -304,7 +189,6 @@ pub struct Instr(&'static InstrDef);
 
 impl Instr {
     /// Create an [`Instr`] referring to the given [`InstrDef`]
-    #[cfg(feature = "microcode")]
     pub(super) fn new(def: &'static InstrDef) -> Self {
         Self(def)
     }
@@ -322,7 +206,6 @@ impl Instr {
     }
 
     /// Retrieve the microcode instruction for the CPU `[Internal Fetch]` operation.
-    #[cfg(feature = "microcode")]
     pub fn internal_fetch() -> Instr {
         /// Static for the CPU's internal fetch instruction.
         static INTERNAL_FETCH: Lazy<InstrDef> = Lazy::new(|| {
@@ -355,38 +238,6 @@ pub enum Reg8 {
     E,
     H,
     L,
-}
-
-impl Reg8 {
-    /// Read this register into the microcode stack.
-    #[cfg(feature = "microcode")]
-    fn read(self, ctx: &mut impl CpuContext) {
-        let val = match self {
-            Reg8::Acc => ctx.cpu().regs.acc,
-            Reg8::B => ctx.cpu().regs.b,
-            Reg8::C => ctx.cpu().regs.c,
-            Reg8::D => ctx.cpu().regs.d,
-            Reg8::E => ctx.cpu().regs.e,
-            Reg8::H => ctx.cpu().regs.h,
-            Reg8::L => ctx.cpu().regs.l,
-        };
-        ctx.cpu_mut().microcode_stack.pushu8(val);
-    }
-
-    /// Write this register from the microcode stack.
-    #[cfg(feature = "microcode")]
-    fn write(self, ctx: &mut impl CpuContext) {
-        let val = ctx.cpu_mut().microcode_stack.popu8();
-        match self {
-            Reg8::Acc => ctx.cpu_mut().regs.acc = val,
-            Reg8::B => ctx.cpu_mut().regs.b = val,
-            Reg8::C => ctx.cpu_mut().regs.c = val,
-            Reg8::D => ctx.cpu_mut().regs.d = val,
-            Reg8::E => ctx.cpu_mut().regs.e = val,
-            Reg8::H => ctx.cpu_mut().regs.h = val,
-            Reg8::L => ctx.cpu_mut().regs.l = val,
-        }
-    }
 }
 
 impl MicrocodeReadable for Reg8 {
@@ -463,34 +314,6 @@ pub enum Reg16 {
     HL,
     Sp,
     Pc,
-}
-
-impl Reg16 {
-    #[cfg(feature = "microcode")]
-    fn read(self, ctx: &mut impl CpuContext) {
-        let val = match self {
-            Reg16::AF => ctx.cpu().regs.af(),
-            Reg16::BC => ctx.cpu().regs.bc(),
-            Reg16::DE => ctx.cpu().regs.de(),
-            Reg16::HL => ctx.cpu().regs.hl(),
-            Reg16::Sp => ctx.cpu().regs.sp,
-            Reg16::Pc => ctx.cpu().regs.pc,
-        };
-        ctx.cpu_mut().microcode_stack.pushu16(val);
-    }
-
-    #[cfg(feature = "microcode")]
-    fn write(self, ctx: &mut impl CpuContext) {
-        let val = ctx.cpu_mut().microcode_stack.popu16();
-        match self {
-            Reg16::AF => ctx.cpu_mut().regs.set_af(val),
-            Reg16::BC => ctx.cpu_mut().regs.set_bc(val),
-            Reg16::DE => ctx.cpu_mut().regs.set_de(val),
-            Reg16::HL => ctx.cpu_mut().regs.set_hl(val),
-            Reg16::Sp => ctx.cpu_mut().regs.sp = val,
-            Reg16::Pc => ctx.cpu_mut().regs.pc = val,
-        }
-    }
 }
 
 impl MicrocodeReadable for Reg16 {
@@ -828,255 +651,6 @@ pub enum Microcode {
     Discard(usize),
 }
 
-impl Microcode {
-    /// Convert a single microcode instruction into a single-instruction builder.
-    pub fn to_builder(self) -> MicrocodeBuilder {
-        MicrocodeBuilder::first(self)
-    }
-
-    #[cfg(feature = "microcode")]
-    pub(super) fn eval(self, ctx: &mut impl CpuContext) -> MicrocodeFlow {
-        match self {
-            Self::Yield => return MicrocodeFlow::Yield1m,
-            Self::ReadReg8(reg) => reg.read(ctx),
-            Self::WriteReg8(reg) => reg.write(ctx),
-            Self::ReadMem8 => ops::read_mem8(ctx),
-            Self::WriteMem8 => ops::write_mem8(ctx),
-            Self::Append(val) => ctx.cpu_mut().microcode_stack.pushu8(val),
-            Self::GetFlagsMasked { mask } => ops::get_flags_masked(ctx, mask),
-            Self::SetFlagsMasked { mask } => ops::set_flags_masked(ctx, mask),
-            Self::BinaryOp(op) => op.eval(ctx),
-            Self::UnaryOp(op) => op.eval(ctx),
-            Self::BitOp(op) => op.eval(ctx),
-            Self::ReadReg16(reg) => reg.read(ctx),
-            Self::WriteReg16(reg) => reg.write(ctx),
-            Self::Inc16 => ops::inc16(ctx),
-            Self::Dec16 => ops::dec16(ctx),
-            Self::Add16 => ops::add16(ctx),
-            Self::OffsetAddr => ops::offset_addr(ctx),
-            Self::Stop => panic!("STOP is bizarre and complicated and not implemented."),
-            Self::Halt => halt(ctx),
-            Self::EnableInterrupts { immediate } => ops::enable_interrupts(ctx, immediate),
-            Self::DisableInterrupts => ctx.cpu_mut().interrupt_master_enable.clear(),
-            Self::CheckHalt => ops::check_halt(ctx),
-            Self::ClearHalt => ctx.cpu_mut().halted = false,
-            Self::CheckIme => ops::check_ime(ctx),
-            Self::GetActiveInterrupts => ops::get_active_interrupts(ctx),
-            Self::PopHaltBug => ops::pop_halt_bug(ctx),
-            Self::PopInterrupt => ops::pop_interrupt(ctx),
-            Self::TickImeOnEnd => ops::tick_ime_on_end(ctx),
-            Self::Skip { steps } => ops::skip(ctx, steps),
-            Self::SkipIf { steps } => ops::skip_if(ctx, steps),
-            Self::FetchNextInstruction => ops::fetch_next_instruction(ctx),
-            Self::ParseOpcode => ops::parse_opcode(ctx),
-            Self::ParseCBOpcode => ops::parse_cb_opcode(ctx),
-            Self::Dup(count) => ctx.cpu_mut().microcode_stack.ustack_dup(count),
-            Self::Swap { top, second } => ctx.cpu_mut().microcode_stack.ustack_swap(top, second),
-            Self::Discard(count) => ctx.cpu_mut().microcode_stack.ustack_discard(count),
-        }
-        MicrocodeFlow::Continue
-    }
-}
-
-/// Contains implementations of microcode operations. Only needed when in microcode mode.
-#[cfg(feature = "microcode")]
-mod ops {
-    use std::mem;
-
-    use super::Instr;
-    use crate::gbz80core::{CBOpcode, CpuContext, Flags, Opcode};
-    use crate::interrupts::Interrupts;
-    use crate::memdev::{Addr, MemDevice};
-
-    /// Pop a 16 bit address and use it to read an 8 bit value from memory onto the stack.
-    pub(super) fn read_mem8(ctx: &mut impl CpuContext) {
-        let addr = Addr::from(ctx.cpu_mut().microcode_stack.popu16());
-        let val = ctx.mem().read(addr);
-        ctx.cpu_mut().microcode_stack.pushu8(val);
-    }
-
-    /// Pop a 16 bit address and an 8 bit value and write the value to the address.
-    pub(super) fn write_mem8(ctx: &mut impl CpuContext) {
-        let addr = Addr::from(ctx.cpu_mut().microcode_stack.popu16());
-        let val = ctx.cpu_mut().microcode_stack.popu8();
-        ctx.mem_mut().write(addr, val);
-    }
-
-    /// Internal implementation of microcode to fetch masked flags to the output.
-    pub(super) fn get_flags_masked(ctx: &mut impl CpuContext, mask: Flags) {
-        let flags = ctx.cpu().regs.flags.intersection(mask);
-        ctx.cpu_mut().microcode_stack.pushu8(flags.bits());
-    }
-
-    /// Internal implementation of microcode to apply masked flags to the output.
-    pub(super) fn set_flags_masked(ctx: &mut impl CpuContext, mask: Flags) {
-        let flags = Flags::from_bits_truncate(ctx.cpu_mut().microcode_stack.popu8());
-        ctx.cpu_mut().regs.flags.merge(flags, mask);
-    }
-
-    /// Internal implementation of microcode u16 increment.
-    pub(super) fn inc16(ctx: &mut impl CpuContext) {
-        let val = ctx.cpu_mut().microcode_stack.popu16();
-        ctx.cpu_mut().microcode_stack.pushu16(val.wrapping_add(1));
-    }
-
-    /// Internal implementation of microcode u16 decrement.
-    pub(super) fn dec16(ctx: &mut impl CpuContext) {
-        let val = ctx.cpu_mut().microcode_stack.popu16();
-        ctx.cpu_mut().microcode_stack.pushu16(val.wrapping_sub(1));
-    }
-
-    /// Internal implementation of microcode u16 add.
-    pub(super) fn add16(ctx: &mut impl CpuContext) {
-        let lhs = ctx.cpu_mut().microcode_stack.popu16();
-        let rhs = ctx.cpu_mut().microcode_stack.popu16();
-
-        let mut flags = Flags::empty();
-        if (lhs & 0x7ff) + (rhs & 0x7ff) > 0x7ff {
-            flags |= Flags::HALFCARRY;
-        }
-        let (res, carry) = lhs.overflowing_add(rhs);
-        flags |= Flags::check_carry(carry);
-
-        ctx.cpu_mut().microcode_stack.pushu16(res);
-        ctx.cpu_mut().microcode_stack.pushu8(flags.bits());
-    }
-
-    pub(super) fn offset_addr(ctx: &mut impl CpuContext) {
-        use crate::gbz80core::oputils::offset_addr as offset_internal;
-        let cpu = ctx.cpu_mut();
-
-        let addr = cpu.microcode_stack.popu16();
-        let offset = cpu.microcode_stack.popu8() as i8;
-
-        let (res, flags) = offset_internal(addr, offset);
-        cpu.microcode_stack.pushu16(res);
-        cpu.microcode_stack.pushu8(flags.bits());
-    }
-
-    /// Internal implementation of the microcode enable interrupts instruction. Runs either
-    /// set or set_next_instruction, depending on whether the interrupt is to be set
-    /// immeidately or not.
-    pub(super) fn enable_interrupts(ctx: &mut impl CpuContext, immediate: bool) {
-        if immediate {
-            ctx.cpu_mut().interrupt_master_enable.set();
-        } else {
-            ctx.cpu_mut().interrupt_master_enable.set_next_instruction();
-        }
-    }
-
-    /// Pushes the value of whether the CPU is halted onto the microcode stack as a u8.
-    pub(super) fn check_halt(ctx: &mut impl CpuContext) {
-        let halted = ctx.cpu().halted as u8;
-        ctx.cpu_mut().microcode_stack.pushu8(halted);
-    }
-
-    /// Pushes the value of whether the CPU IME is enabled onto the microcode stack as a
-    /// u8.
-    pub(super) fn check_ime(ctx: &mut impl CpuContext) {
-        let ime = ctx.cpu().interrupt_master_enable.enabled() as u8;
-        ctx.cpu_mut().microcode_stack.pushu8(ime);
-    }
-
-    /// Pushes the set of interrupts which are both active and enabled onto the microcode
-    /// stack.
-    pub(super) fn get_active_interrupts(ctx: &mut impl CpuContext) {
-        let active = ctx.interrupts().active().bits();
-        ctx.cpu_mut().microcode_stack.pushu8(active);
-    }
-
-    /// Pushes the value of the halt_bug flag onto the microcode stack, clearing the value.
-    pub(super) fn pop_halt_bug(ctx: &mut impl CpuContext) {
-        let halt_bug = mem::replace(&mut ctx.cpu_mut().halt_bug, false) as u8;
-        ctx.cpu_mut().microcode_stack.pushu8(halt_bug);
-    }
-
-    /// Pushes the destination address of the next active and enabled interrupt onto the
-    /// microcode stack and clears that interrupt. Panics if no interrupts are active!.
-    pub(super) fn pop_interrupt(ctx: &mut impl CpuContext) {
-        match ctx.interrupts().active().iter().next() {
-            Some(interrupt) => {
-                ctx.interrupts_mut().clear(interrupt);
-                ctx.cpu_mut().microcode_stack.pushu16(interrupt.handler_addr());
-            }
-            None => panic!("Must not use the PopInterrupt microcode instruction if there are no active interrupts."),
-        }
-    }
-
-    /// Enables IME ticking on the next `FetchNextInstruction`.
-    pub(super) fn tick_ime_on_end(ctx: &mut impl CpuContext) {
-        let cpu = ctx.cpu_mut();
-        debug_assert!(
-            cpu.prev_ime.is_none(),
-            "prev_ime is already set to {:?}, trying to set {:?}",
-            cpu.prev_ime.unwrap(),
-            cpu.interrupt_master_enable
-        );
-        cpu.prev_ime = Some(cpu.interrupt_master_enable);
-    }
-
-    /// Skip the CPU forward by this number of microcode instruction steps.
-    pub(super) fn skip(ctx: &mut impl CpuContext, steps: usize) {
-        // Only checked for overflow in debug.
-        ctx.cpu_mut().microcode_pc += steps;
-        debug_assert!(ctx.cpu().microcode_pc <= ctx.cpu().instruction.len());
-    }
-
-    /// Skip the CPU forward by this number of microcode instruction steps if the u8 value on
-    /// top of the microcode stack is non-zero.
-    pub(super) fn skip_if(ctx: &mut impl CpuContext, steps: usize) {
-        let cond = ctx.cpu_mut().microcode_stack.popu8();
-        if cond != 0 {
-            // Only checked for overflow in debug.
-            ctx.cpu_mut().microcode_pc += steps;
-            debug_assert!(ctx.cpu().microcode_pc <= ctx.cpu().instruction.len());
-        }
-    }
-
-    /// Resets to the CPU Internal instruction.
-    pub(super) fn fetch_next_instruction(ctx: &mut impl CpuContext) {
-        let cpu = ctx.cpu_mut();
-        debug_assert!(
-            cpu.microcode_stack.is_empty(),
-            "Previous Instruction {} failed to empty its stack",
-            cpu.instruction.label()
-        );
-        if let Some(prev_ime) = cpu.prev_ime.take() {
-            cpu.interrupt_master_enable.tick(prev_ime);
-        }
-        cpu.microcode_pc = 0;
-        cpu.instruction = Instr::internal_fetch();
-    }
-
-    /// Parses a regular opcode from the microcode stack and replaces the current
-    /// instruction with it.
-    pub(super) fn parse_opcode(ctx: &mut impl CpuContext) {
-        let cpu = ctx.cpu_mut();
-        let opcode = cpu.microcode_stack.popu8();
-        debug_assert!(
-            cpu.microcode_stack.is_empty(),
-            "Previous Instruction {} failed to empty its stack",
-            cpu.instruction.label()
-        );
-        cpu.microcode_pc = 0;
-        cpu.instruction = Opcode::get_instruction(opcode);
-    }
-
-    /// Parses a cb opcode from the microcode stack and replaces the current
-    /// instruction with it.
-    pub(super) fn parse_cb_opcode(ctx: &mut impl CpuContext) {
-        let cpu = ctx.cpu_mut();
-        let opcode = cpu.microcode_stack.popu8();
-        debug_assert!(
-            cpu.microcode_stack.is_empty(),
-            "Previous Instruction {} failed to empty its stack",
-            cpu.instruction.label()
-        );
-        cpu.microcode_pc = 0;
-        cpu.instruction = CBOpcode::get_instruction(opcode);
-    }
-}
-
 /// These are the binary operations available in the microcode. They largely correspond to
 /// operations available in the ALU on the gameboy. All of these operations pop two values
 /// off of the microcode stack and push their result followed by their resulting flags
@@ -1098,57 +672,6 @@ pub enum BinaryOp {
     And,
     Xor,
     Or,
-}
-
-impl BinaryOp {
-    #[cfg(feature = "microcode")]
-    fn eval(self, ctx: &mut impl CpuContext) {
-        let cpu = ctx.cpu_mut();
-        let lhs = cpu.microcode_stack.popu8();
-        let rhs = cpu.microcode_stack.popu8();
-        let (res, flags) = match self {
-            Self::Add => add8_flags(lhs, rhs),
-            Self::AddCarry => {
-                let (mut res, mut flags) = add8_flags(lhs, rhs);
-                if cpu.regs.flags.contains(Flags::CARRY) {
-                    let (res2, flags2) = add8_flags(res, 1);
-                    res = res2;
-                    // Zero flag should only be set if the second add had a result of zero.
-                    flags = flags2 | (flags - Flags::ZERO);
-                }
-                (res, flags)
-            }
-            Self::Sub => sub8_flags(lhs, rhs),
-            Self::SubCarry => {
-                let (mut res, mut flags) = sub8_flags(lhs, rhs);
-                if cpu.regs.flags.contains(Flags::CARRY) {
-                    let (res2, flags2) = sub8_flags(res, 1);
-                    res = res2;
-                    // Zero flag should only be set if the second subtract had a result of
-                    // zero.
-                    flags = flags2 | (flags - Flags::ZERO);
-                }
-                (res, flags)
-            }
-            Self::And => {
-                let res = lhs & rhs;
-                let flags = Flags::HALFCARRY | Flags::check_zero(res);
-                (res, flags)
-            }
-            Self::Or => {
-                let res = lhs | rhs;
-                let flags = Flags::check_zero(res);
-                (res, flags)
-            }
-            Self::Xor => {
-                let res = lhs ^ rhs;
-                let flags = Flags::check_zero(res);
-                (res, flags)
-            }
-        };
-        cpu.microcode_stack.pushu8(res);
-        cpu.microcode_stack.pushu8(flags.bits);
-    }
 }
 
 impl From<BinaryOp> for MicrocodeBuilder {
@@ -1199,79 +722,6 @@ pub enum UnaryOp {
     Swap,
 }
 
-impl UnaryOp {
-    #[cfg(feature = "microcode")]
-    fn eval(self, ctx: &mut impl CpuContext) {
-        let val = ctx.cpu_mut().microcode_stack.popu8();
-        let (res, flags) = match self {
-            Self::RotateLeft8 => {
-                let res = val.rotate_left(1);
-                let flags = Flags::check_zero(res) | Flags::check_carry(res & 1 != 0);
-                (res, flags)
-            }
-            Self::RotateLeft9 => rotate_left9(val, ctx.cpu().regs.flags),
-            Self::RotateRight8 => {
-                let res = val.rotate_right(1);
-                let flags = Flags::check_zero(res) | Flags::check_carry(res & 0x80 != 0);
-                (res, flags)
-            }
-            Self::RotateRight9 => rotate_right9(val, ctx.cpu().regs.flags),
-            Self::DecimalAdjust => {
-                let inflags = ctx.cpu().regs.flags;
-                // Always clears HALFCARRY.
-                let mut flags = Flags::empty();
-                let mut res = val;
-                if inflags.contains(Flags::SUB) {
-                    if inflags.contains(Flags::CARRY) {
-                        flags |= Flags::CARRY;
-                        res = res.wrapping_sub(0x60);
-                    }
-                    if inflags.contains(Flags::HALFCARRY) {
-                        res = res.wrapping_sub(0x06);
-                    }
-                } else {
-                    if inflags.contains(Flags::CARRY) || val > 0x99 {
-                        flags |= Flags::CARRY;
-                        res = res.wrapping_add(0x60);
-                    }
-                    if inflags.contains(Flags::HALFCARRY) || res & 0xf > 9 {
-                        res = res.wrapping_add(0x06);
-                    }
-                }
-                flags |= Flags::check_zero(res);
-                (res, flags)
-            }
-            Self::Compliment => {
-                const FLAGS: Flags = Flags::SUB.union(Flags::HALFCARRY);
-                let res = !val;
-                (res, FLAGS)
-            }
-            Self::ShiftLeft => {
-                let res = val << 1;
-                let flags = Flags::check_zero(res) | Flags::check_carry(val & 0x80 != 0);
-                (res, flags)
-            }
-            Self::ShiftRight => {
-                let res = val >> 1;
-                let flags = Flags::check_zero(res) | Flags::check_carry(val & 1 != 0);
-                (res, flags)
-            }
-            Self::ShiftRightSignExt => {
-                let res = ((val as i8) >> 1) as u8;
-                let flags = Flags::check_zero(res) | Flags::check_carry(val & 1 != 0);
-                (res, flags)
-            }
-            Self::Swap => {
-                let res = ((val & 0x0f) << 4) | ((val & 0xf0) >> 4);
-                let flags = Flags::check_zero(res);
-                (res, flags)
-            }
-        };
-        ctx.cpu_mut().microcode_stack.pushu8(res);
-        ctx.cpu_mut().microcode_stack.pushu8(flags.bits());
-    }
-}
-
 impl From<UnaryOp> for MicrocodeBuilder {
     fn from(op: UnaryOp) -> Self {
         Microcode::UnaryOp(op).into()
@@ -1291,22 +741,6 @@ pub enum BitOp {
     SetBit(u8),
     /// Clears the given bit (given as an index in range 0..=7) in the operand.
     ResetBit(u8),
-}
-
-impl BitOp {
-    #[cfg(feature = "microcode")]
-    fn eval(self, ctx: &mut impl CpuContext) {
-        let val = ctx.cpu_mut().microcode_stack.popu8();
-        let res = match self {
-            Self::TestBit(bit) => {
-                let flags = Flags::check_zero(val & (1 << bit)) | Flags::HALFCARRY;
-                flags.bits()
-            }
-            Self::SetBit(bit) => val | (1 << bit),
-            Self::ResetBit(bit) => val & !(1 << bit),
-        };
-        ctx.cpu_mut().microcode_stack.pushu8(res);
-    }
 }
 
 impl From<BitOp> for MicrocodeBuilder {
